+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     The readers and writers of the inner type are used instead of the ones for products.
   - `ConfigReader` and `ConfigWriter` instances for `Pattern` and `Regex`.
   - `ConfigReader` and `ConfigWriter` for `java.io.File`
+  - `ConfigReader` and `ConfigWriter` for arbitrary Java Enums
 - Bug fixes
   - `Duration.Undefined` is correctly handled when reading and writing configurations. [[#184](https://github.com/melrief/pureconfig/issues/184)]